// src/api/generated/models/questionResource.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * brew
 * OpenAPI spec version: 0.0.1
 */
import type { QuestionResourceQuestionPt } from './questionResourceQuestionPt';
import type { QuestionResourceQuestionEn } from './questionResourceQuestionEn';
import type { QuestionResourceLanguageId } from './questionResourceLanguageId';
import type { QuestionResourceCodeSnippet } from './questionResourceCodeSnippet';
import type { QuestionResponseResource } from './questionResponseResource';

export interface QuestionResource {
  id: string;
  question: string;
  question_pt: QuestionResourceQuestionPt;
  question_en: QuestionResourceQuestionEn;
  translation_status: string;
  difficulty_level: number;
  language_id: QuestionResourceLanguageId;
  category: string;
  ideal_time_to_solve: number;
  code_snippet: QuestionResourceCodeSnippet;
  is_multiple_choice: boolean;
  seniority_level: string;
  responses: QuestionResponseResource[];
}